+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Copyright (c) 2022-2026 Claudio Satriano <satriano@ipgp.fr>
 - New plottype `plotly` for `seiscat plot`, producing interactive 3D seismicity
   plots
 - New command `seiscat logo` to print the beautiful, ascii-art SeisCat logo
+- New option for `seiscat plot`: `--out-file` to save the plot to a file
 - Fixed an issue in `seiscat editdb` where positional arguments (`eventid` and
   `event_version`) were not parsed correctly due to improper handling of the
   `--set` and `--increment` options. Users must now repeat the `--set` and